minor updates to Makefile and parsing for server.c

+++ trunk/p5/output.c
@@ -3,7 +3,6 @@
 #include <assert.h>
 #include <unistd.h>
 
-
 //
 // This program is intended to help you test your web server.
 // You can use it to test that you are correctly having multiple threads
@@ -45,13 +44,14 @@ int main(int argc, char *argv[])
       sleep(1);
   double t2 = Time_GetSeconds();
 
+
   /* Make the response body */
   sprintf(content, "<p>Welcome to the CGI program</p>\r\n");
   sprintf(content, "%s<p>My only purpose is to waste time on the server!</p>\r\n", content);
   sprintf(content, "%s<p>I spun for %.2f seconds</p>\r\n", content, t2 - t1);
   
   /* Generate the HTTP response */
-  printf("Content-length: %d\r\n", strlen(content));
+  printf("Content-length: %d\r\n", (int)strlen(content));
   printf("Content-type: text/html\r\n\r\n");
   printf("%s", content);
   fflush(stdout);
